+++ nix/pkgs/pinpon-hlint.nix
@@ -3,9 +3,9 @@
 , hlint, hpio, hspec, http-client, http-client-tls, http-types
 , lens, lucid, mtl, network, optparse-applicative, optparse-text
 , protolude, QuickCheck, quickcheck-instances, resourcet, servant
-, servant-client, servant-docs, servant-lucid, servant-server
-, servant-swagger, servant-swagger-ui, stdenv, swagger2, text, time
-, transformers, transformers-base, wai, warp
+, servant-client, servant-client-core, servant-docs, servant-lucid
+, servant-server, servant-swagger, servant-swagger-ui, stdenv
+, swagger2, text, time, transformers, transformers-base, wai, warp
 }:
 mkDerivation {
   pname = "pinpon";
@@ -18,15 +18,16 @@ mkDerivation {
   libraryHaskellDepends = [
     aeson aeson-pretty amazonka amazonka-core amazonka-sns base
     bytestring containers exceptions http-client http-types lens lucid
-    mtl protolude resourcet servant servant-client servant-docs
-    servant-lucid servant-server servant-swagger servant-swagger-ui
-    swagger2 text time transformers transformers-base wai warp
+    mtl protolude resourcet servant servant-client servant-client-core
+    servant-docs servant-lucid servant-server servant-swagger
+    servant-swagger-ui swagger2 text time transformers
+    transformers-base wai warp
   ];
   executableHaskellDepends = [
     amazonka amazonka-sns base bytestring containers exceptions hpio
     http-client http-client-tls http-types lens mtl network
-    optparse-applicative optparse-text protolude servant-client text
-    time transformers warp
+    optparse-applicative optparse-text protolude servant-client
+    servant-client-core text time transformers warp
   ];
   testHaskellDepends = [
     aeson base bytestring doctest exceptions hlint hspec protolude